+++ src/app/web5/todolist-shared/components/NewListCreator.tsx
@@ -1,4 +1,4 @@
-export const CreateNewTodoList = () => {
+export const NewListCreator = () => {
 	return (
 		<div>
 			<h1>Create New Todo List</h1>
