+++ src/app/web5/todolist-shared/constants.ts
@@ -3,11 +3,6 @@
 // and host it somewhere like S3, google cloud storage, or github
 export const TODO_SCHEMA = "https://schema.org/ToDo";
 
-export interface Web5ContextState {
-	web5: any;
-	did: string;
-	updateProtocol: (protocolDefinition: any) => void;
-}
 export interface Todo {
 	record: any;
 	data: {
@@ -15,17 +10,6 @@ export interface Todo {
 		description: string;
 	};
 	id: string;
-}
-
-export interface TodoDwnContextState {
-	todos: Todo[];
-	addTodo: any;
-	deleteTodo: any;
-	updateTodo: any;
-}
-export interface ListsContextState {
-	lists: TodoList[];
-	addList: any;
 }
 
 export interface NewTodoData {
