+++ src/app/web5/todolist-shared/page.tsx
@@ -1,12 +1,11 @@
 "use client";
 
-import { SharedTodoList } from "./SharedTodoList";
+import { List } from "./components/List";
 import { BasePage } from "@/app/shared/BasePage";
 import { useEffect, useState } from "react";
-import { TodoDwnProvider } from "./TodoDwnProvider";
-import { DIDViewer } from "./DIDViewer";
-import { Web5 } from "@web5/api";
-import { Web5Provider } from "./Web5Provider";
+import { TodoProvider } from "./providers/TodoProvider";
+import { DIDViewer } from "./components/DIDViewer";
+import { Web5Provider } from "./providers/Web5Provider";
 
 // `app/page.tsx` is the UI for the `/` URL
 
@@ -42,12 +41,12 @@ export default function Page() {
 	console.log("data", data);
 	return (
 		<Web5Provider protocolDefinition={data}>
-			<TodoDwnProvider protocolDefinition={data}>
-				<BasePage title="Web5 Shared Todo List" description={PAGE_DESCRIPTION}>
-					<DIDViewer />
-					<SharedTodoList />
-				</BasePage>
-			</TodoDwnProvider>
+			<BasePage title="Web5 Shared Todo List" description={PAGE_DESCRIPTION}>
+				<DIDViewer />
+				<TodoProvider protocolDefinition={data}>
+					<List />
+				</TodoProvider>
+			</BasePage>
 		</Web5Provider>
 	);
 }
