+++ src/app/web5/todolist-shared/providers/ListsProvider.tsx
@@ -7,15 +7,20 @@ import {
 	useState,
 } from "react";
 import { createListRecord, retrieveLists } from "./utils";
-import { ListsContextState, Todo, TodoList } from "./constants";
+import { TodoList } from "../constants";
 import { useWeb5 } from "./Web5Provider";
 
-const TodoDwnContext = createContext<ListsContextState>({
+export interface ListsContextState {
+	lists: TodoList[];
+	addList: any;
+}
+
+const ListsContext = createContext<ListsContextState>({
 	lists: [],
 	addList: () => {},
 });
 
-export const ListDwnProvider = ({ children, protocolDefinition }: any) => {
+export const ListsProvider = ({ children, protocolDefinition }: any) => {
 	const [lists, setLists] = useState<TodoList[]>([]);
 	const { web5, did } = useWeb5();
 	useEffect(() => {
@@ -80,12 +85,12 @@ export const ListDwnProvider = ({ children, protocolDefinition }: any) => {
 	);
 
 	return (
-		<TodoDwnContext.Provider value={value}>{children}</TodoDwnContext.Provider>
+		<ListsContext.Provider value={value}>{children}</ListsContext.Provider>
 	);
 };
 
 export const useTodoListManager = () => {
-	const context = useContext(TodoDwnContext);
+	const context = useContext(ListsContext);
 	if (!context) {
 		throw new Error("useTodoListManager must be used within a Web5Provider");
 	}
